+++ routes/web.php
@@ -157,13 +157,6 @@ function()
             }
 
             $type = (Request::input('format') == 'digitálna reprodukcia') ? 'digitálna' : 'tlačená';
-            $purpose = ($order->purpose) ? $order->purpose : $order->frame;
-
-            // @TODO: remove this after EEA fix accepting empty purpose
-            if (empty($purpose)) {
-                $purpose =  'účel';
-            }
-            // /@TODO
 
             //poslat objednavku do Jiry
             $client = new GuzzleHttp\Client();
@@ -175,7 +168,7 @@ function()
                     'contactPerson' => $order->name,
                     'email' => $order->email,
                     'kindOfPurpose' => $order->purpose_kind,
-                    'purpose' => $purpose,
+                    'purpose' => $order->purpose,
                     'medium' => 'Iné',
                     'address' => $order->address,
                     'phone' => $order->phone,
@@ -184,6 +177,7 @@ function()
                     'numOfCopies' => '1',
                     'reproductionType' => $type,
                     'format' => $order->format,
+                    'frameColor' => $order->frame,
                     'deliveryPoint' => $order->delivery_point,
                     'note' => $order->note,
                 ],